move useModesMessage and displayModesMessage to net_io. move interesting static functions from modes.c to readsb.c and expose in header
