+++ mode_s.c
@@ -53,6 +53,7 @@
 
 
 #include "readsb.h"
+#include "mode_s.h"
 #include "ais_charset.h"
 
 /* for PRIX64 */
@@ -63,11 +64,6 @@
 //
 //
 //
-
-/* A timestamp that indicates the data is synthetic, created from a
- * multilateration result
- */
-#define MAGIC_MLAT_TIMESTAMP 0xFF004D4C4154ULL
 
 //=========================================================================
 //
@@ -1473,209 +1469,7 @@ static void decodeExtendedSquitter(struct modesMessage *mm) {
     }
 }
 
-static const char *df_names[33] = {
-    /* 0 */ "Short Air-Air Surveillance",
-    /* 1 */ NULL,
-    /* 2 */ NULL,
-    /* 3 */ NULL,
-    /* 4 */ "Survelliance, Altitude Reply",
-    /* 5 */ "Survelliance, Identity Reply",
-    /* 6 */ NULL,
-    /* 7 */ NULL,
-    /* 8 */ NULL,
-    /* 9 */ NULL,
-    /* 10 */ NULL,
-    /* 11 */ "All Call Reply",
-    /* 12 */ NULL,
-    /* 13 */ NULL,
-    /* 14 */ NULL,
-    /* 15 */ NULL,
-    /* 16 */ "Long Air-Air ACAS",
-    /* 17 */ "Extended Squitter",
-    /* 18 */ "Extended Squitter (Non-Transponder)",
-    /* 19 */ "Extended Squitter (Military)",
-    /* 20 */ "Comm-B, Altitude Reply",
-    /* 21 */ "Comm-B, Identity Reply",
-    /* 22 */ "Military Use",
-    /* 23 */ NULL,
-    /* 24 */ "Comm-D Extended Length Message",
-    /* 25 */ "Comm-D Extended Length Message",
-    /* 26 */ "Comm-D Extended Length Message",
-    /* 27 */ "Comm-D Extended Length Message",
-    /* 28 */ "Comm-D Extended Length Message",
-    /* 29 */ "Comm-D Extended Length Message",
-    /* 30 */ "Comm-D Extended Length Message",
-    /* 31 */ "Comm-D Extended Length Message",
-    /* 32 */ "Mode A/C Reply",
-};
-
-static const char *df_to_string(unsigned df) {
-    if (df > 32)
-        return "out of range";
-    if (!df_names[df])
-        return "reserved";
-    return df_names[df];
-}
-
-static const char *altitude_unit_to_string(altitude_unit_t unit) {
-    switch (unit) {
-        case UNIT_FEET:
-            return "ft";
-        case UNIT_METERS:
-            return "m";
-        default:
-            return "(unknown altitude unit)";
-    }
-}
-
-static const char *airground_to_string(AircraftMeta__AirGround airground) {
-    switch (airground) {
-        case AIRCRAFT_META__AIR_GROUND__AG_GROUND:
-            return "ground";
-        case AIRCRAFT_META__AIR_GROUND__AG_AIRBORNE:
-            return "airborne";
-        case AIRCRAFT_META__AIR_GROUND__AG_INVALID:
-            return "invalid";
-        case AIRCRAFT_META__AIR_GROUND__AG_UNCERTAIN:
-            return "airborne?";
-        default:
-            return "(unknown airground state)";
-    }
-}
-
-static const char *addrtype_to_string(AircraftMeta__AddrType type) {
-    switch (type) {
-        case AIRCRAFT_META__ADDR_TYPE__ADDR_ADSB_ICAO:
-            return "Mode S / ADS-B";
-        case AIRCRAFT_META__ADDR_TYPE__ADDR_ADSB_ICAO_NT:
-            return "ADS-B, non-transponder";
-        case AIRCRAFT_META__ADDR_TYPE__ADDR_ADSB_OTHER:
-            return "ADS-B, other addressing scheme";
-        case AIRCRAFT_META__ADDR_TYPE__ADDR_TISB_ICAO:
-            return "TIS-B";
-        case AIRCRAFT_META__ADDR_TYPE__ADDR_TISB_OTHER:
-            return "TIS-B, other addressing scheme";
-        case AIRCRAFT_META__ADDR_TYPE__ADDR_TISB_TRACKFILE:
-            return "TIS-B, Mode A code and track file number";
-        case AIRCRAFT_META__ADDR_TYPE__ADDR_ADSR_ICAO:
-            return "ADS-R";
-        case AIRCRAFT_META__ADDR_TYPE__ADDR_ADSR_OTHER:
-            return "ADS-R, other addressing scheme";
-        case AIRCRAFT_META__ADDR_TYPE__ADDR_MODE_A:
-            return "Mode A";
-        default:
-            return "unknown addressing scheme";
-    }
-}
-
-static const char *cpr_type_to_string(cpr_type_t type) {
-    switch (type) {
-        case CPR_SURFACE:
-            return "Surface";
-        case CPR_AIRBORNE:
-            return "Airborne";
-        case CPR_COARSE:
-            return "TIS-B Coarse";
-        default:
-            return "unknown CPR type";
-    }
-}
-
-static const char *heading_type_to_string(heading_type_t type) {
-    switch (type) {
-        case HEADING_GROUND_TRACK:
-            return "Ground track";
-        case HEADING_MAGNETIC:
-            return "Mag heading";
-        case HEADING_TRUE:
-            return "True heading";
-        case HEADING_MAGNETIC_OR_TRUE:
-            return "Heading";
-        case HEADING_TRACK_OR_HEADING:
-            return "Track/Heading";
-        default:
-            return "unknown heading type";
-    }
-}
-
-static const char *commb_format_to_string(commb_format_t format) {
-    switch (format) {
-        case COMMB_EMPTY_RESPONSE:
-            return "empty response";
-        case COMMB_AMBIGUOUS:
-            return "ambiguous format";
-        case COMMB_DATALINK_CAPS:
-            return "BDS1,0 Datalink capabilities";
-        case COMMB_GICB_CAPS:
-            return "BDS1,7 Common usage GICB capabilities";
-        case COMMB_AIRCRAFT_IDENT:
-            return "BDS2,0 Aircraft identification";
-        case COMMB_ACAS_RA:
-            return "BDS3,0 ACAS resolution advisory";
-        case COMMB_VERTICAL_INTENT:
-            return "BDS4,0 Selected vertical intention";
-        case COMMB_TRACK_TURN:
-            return "BDS5,0 Track and turn report";
-        case COMMB_HEADING_SPEED:
-            return "BDS6,0 Heading and speed report";
-        default:
-            return "unknown format";
-    }
-}
-
-static const char *nav_modes_to_string(nav_modes_t flags) {
-    static char buf[128];
-
-    buf[0] = 0;
-    if (flags & NAV_MODE_AUTOPILOT)
-        strcat(buf, "autopilot ");
-    if (flags & NAV_MODE_VNAV)
-        strcat(buf, "vnav ");
-    if (flags & NAV_MODE_ALT_HOLD)
-        strcat(buf, "althold ");
-    if (flags & NAV_MODE_APPROACH)
-        strcat(buf, "approach ");
-    if (flags & NAV_MODE_LNAV)
-        strcat(buf, "lnav ");
-    if (flags & NAV_MODE_TCAS)
-        strcat(buf, "tcas ");
-
-    if (buf[0] != 0)
-        buf[strlen(buf) - 1] = 0;
-
-    return buf;
-}
-
-static const char *sil_type_to_string(AircraftMeta__SilType type) {
-    switch (type) {
-        case AIRCRAFT_META__SIL_TYPE__SIL_UNKNOWN: return "unknown type";
-        case AIRCRAFT_META__SIL_TYPE__SIL_PER_HOUR: return "per flight hour";
-        case AIRCRAFT_META__SIL_TYPE__SIL_PER_SAMPLE: return "per sample";
-        default: return "invalid type";
-    }
-}
-
-static const char *emergency_to_string(AircraftMeta__Emergency emergency) {
-    switch (emergency) {
-        case AIRCRAFT_META__EMERGENCY__EMERGENCY_NONE: return "no emergency";
-        case AIRCRAFT_META__EMERGENCY__EMERGENCY_GENERAL: return "general emergency (7700)";
-        case AIRCRAFT_META__EMERGENCY__EMERGENCY_LIFEGUARD: return "lifeguard / medical emergency";
-        case AIRCRAFT_META__EMERGENCY__EMERGENCY_MINFUEL: return "minimum fuel";
-        case AIRCRAFT_META__EMERGENCY__EMERGENCY_NORDO: return "no communications (7600)";
-        case AIRCRAFT_META__EMERGENCY__EMERGENCY_UNLAWFUL: return "unlawful interference (7500)";
-        case AIRCRAFT_META__EMERGENCY__EMERGENCY_DOWNED: return "downed aircraft";
-        default: return "reserved";
-    }
-}
-
-static void print_hex_bytes(unsigned char *data, size_t len) {
-    size_t i;
-    for (i = 0; i < len; ++i) {
-        printf("%02X", (unsigned) data[i]);
-    }
-}
-
-static int esTypeHasSubtype(unsigned metype) {
+int esTypeHasSubtype(unsigned metype) {
     if (metype <= 18) {
         return 0;
     }
@@ -1687,7 +1481,7 @@ static int esTypeHasSubtype(unsigned metype) {
     return 1;
 }
 
-static const char *esTypeName(unsigned metype, unsigned mesub) {
+const char *esTypeName(unsigned metype, unsigned mesub) {
     switch (metype) {
         case 0:
             return "No position information (airborne or surface)";
@@ -1771,408 +1565,6 @@ static const char *esTypeName(unsigned metype, unsigned mesub) {
 
         default:
             return "Unknown";
-    }
-}
-
-void displayModesMessage(struct modesMessage *mm) {
-    int j;
-
-    // Handle only addresses mode first.
-    if (Modes.onlyaddr) {
-        printf("%06x\n", mm->addr);
-        return; // Enough for --onlyaddr mode
-    }
-
-    // Show the raw message.
-    if (Modes.mlat && mm->timestampMsg) {
-        printf("@%012" PRIX64, mm->timestampMsg);
-    } else
-        printf("*");
-
-    for (j = 0; j < mm->msgbits / 8; j++) printf("%02x", mm->msg[j]);
-    printf(";\n");
-
-    if (Modes.raw) {
-        fflush(stdout); // Provide data to the reader ASAP
-        return; // Enough for --raw mode
-    }
-
-    if (mm->msgtype < 32)
-        printf("CRC: %06x\n", mm->crc);
-
-    if (mm->correctedbits != 0)
-        printf("No. of bit errors fixed: %d\n", mm->correctedbits);
-
-    if (mm->signalLevel > 0)
-        printf("RSSI: %.1f dBFS\n", 10 * log10(mm->signalLevel));
-
-    if (mm->score)
-        printf("Score: %d\n", mm->score);
-
-    if (mm->timestampMsg == MAGIC_MLAT_TIMESTAMP)
-        printf("This is a synthetic MLAT message.\n");
-    else
-        printf("Time: %.2fus\n", mm->timestampMsg / 12.0);
-
-    switch (mm->msgtype) {
-        case 0:
-            printf("DF:0 addr:%06X VS:%u CC:%u SL:%u RI:%u AC:%u\n",
-                    mm->addr, mm->VS, mm->CC, mm->SL, mm->RI, mm->AC);
-            break;
-
-        case 4:
-            printf("DF:4 addr:%06X FS:%u DR:%u UM:%u AC:%u\n",
-                    mm->addr, mm->FS, mm->DR, mm->UM, mm->AC);
-            break;
-
-        case 5:
-            printf("DF:5 addr:%06X FS:%u DR:%u UM:%u ID:%u\n",
-                    mm->addr, mm->FS, mm->DR, mm->UM, mm->ID);
-            break;
-
-        case 11:
-            printf("DF:11 AA:%06X IID:%u CA:%u\n",
-                    mm->AA, mm->IID, mm->CA);
-            break;
-
-        case 16:
-            printf("DF:16 addr:%06x VS:%u SL:%u RI:%u AC:%u MV:",
-                    mm->addr, mm->VS, mm->SL, mm->RI, mm->AC);
-            print_hex_bytes(mm->MV, sizeof (mm->MV));
-            printf("\n");
-            break;
-
-        case 17:
-            printf("DF:17 AA:%06X CA:%u ME:",
-                    mm->AA, mm->CA);
-            print_hex_bytes(mm->ME, sizeof (mm->ME));
-            printf("\n");
-            break;
-
-        case 18:
-            printf("DF:18 AA:%06X CF:%u ME:",
-                    mm->AA, mm->CF);
-            print_hex_bytes(mm->ME, sizeof (mm->ME));
-            printf("\n");
-            break;
-
-        case 20:
-            printf("DF:20 addr:%06X FS:%u DR:%u UM:%u AC:%u MB:",
-                    mm->addr, mm->FS, mm->DR, mm->UM, mm->AC);
-            print_hex_bytes(mm->MB, sizeof (mm->MB));
-            printf("\n");
-            break;
-
-        case 21:
-            printf("DF:21 addr:%06x FS:%u DR:%u UM:%u ID:%u MB:",
-                    mm->addr, mm->FS, mm->DR, mm->UM, mm->ID);
-            print_hex_bytes(mm->MB, sizeof (mm->MB));
-            printf("\n");
-            break;
-
-        case 24:
-        case 25:
-        case 26:
-        case 27:
-        case 28:
-        case 29:
-        case 30:
-        case 31:
-            printf("DF:24 addr:%06x KE:%u ND:%u MD:",
-                    mm->addr, mm->KE, mm->ND);
-            print_hex_bytes(mm->MD, sizeof (mm->MD));
-            printf("\n");
-            break;
-    }
-
-    printf(" %s", df_to_string(mm->msgtype));
-    if (mm->msgtype == 17 || mm->msgtype == 18) {
-        if (esTypeHasSubtype(mm->metype)) {
-            printf(" %s (%u/%u)",
-                    esTypeName(mm->metype, mm->mesub),
-                    mm->metype,
-                    mm->mesub);
-        } else {
-            printf(" %s (%u)",
-                    esTypeName(mm->metype, mm->mesub),
-                    mm->metype);
-        }
-    }
-    printf("\n");
-
-    if (mm->msgtype == 20 || mm->msgtype == 21) {
-        printf("  Comm-B format: %s\n", commb_format_to_string(mm->commb_format));
-    }
-
-    if (mm->addr & MODES_NON_ICAO_ADDRESS) {
-        printf("  Other Address: %06X (%s)\n", mm->addr & 0xFFFFFF, addrtype_to_string(mm->addrtype));
-    } else {
-        printf("  ICAO Address:  %06X (%s)\n", mm->addr, addrtype_to_string(mm->addrtype));
-    }
-
-    if (mm->airground != AIRCRAFT_META__AIR_GROUND__AG_INVALID) {
-        printf("  Air/Ground:    %s\n",
-                airground_to_string(mm->airground));
-    }
-
-    if (mm->altitude_baro_valid) {
-        printf("  Baro altitude: %d %s\n",
-                mm->altitude_baro,
-                altitude_unit_to_string(mm->altitude_baro_unit));
-    }
-
-    if (mm->altitude_geom_valid) {
-        printf("  Geom altitude: %d %s\n",
-                mm->altitude_geom,
-                altitude_unit_to_string(mm->altitude_geom_unit));
-    }
-
-    if (mm->geom_delta_valid) {
-        printf("  Geom - baro:   %d ft\n",
-                mm->geom_delta);
-    }
-
-    if (mm->heading_valid) {
-        printf("  %-13s  %.1f\n", heading_type_to_string(mm->heading_type), mm->heading);
-    }
-
-    if (mm->track_rate_valid) {
-        printf("  Track rate:    %.2f deg/sec %s\n", mm->track_rate, mm->track_rate < 0 ? "left" : mm->track_rate > 0 ? "right" : "");
-    }
-
-    if (mm->roll_valid) {
-        printf("  Roll:          %.1f degrees %s\n", mm->roll, mm->roll < -0.05 ? "left" : mm->roll > 0.05 ? "right" : "");
-    }
-
-    if (mm->gs_valid) {
-        printf("  Groundspeed:   %.1f kt", mm->gs.selected);
-        if (mm->gs.v0 != mm->gs.selected) {
-            printf(" (v0: %.1f kt)", mm->gs.v0);
-        }
-        if (mm->gs.v2 != mm->gs.selected) {
-            printf(" (v2: %.1f kt)", mm->gs.v2);
-        }
-        printf("\n");
-    }
-
-    if (mm->ias_valid) {
-        printf("  IAS:           %u kt\n", mm->ias);
-    }
-
-    if (mm->tas_valid) {
-        printf("  TAS:           %u kt\n", mm->tas);
-    }
-
-    if (mm->mach_valid) {
-        printf("  Mach number:   %.3f\n", mm->mach);
-    }
-
-    if (mm->baro_rate_valid) {
-        printf("  Baro rate:     %d ft/min\n", mm->baro_rate);
-    }
-
-    if (mm->geom_rate_valid) {
-        printf("  Geom rate:     %d ft/min\n", mm->geom_rate);
-    }
-
-    if (mm->squawk_valid) {
-        printf("  Squawk:        %04x\n",
-                mm->squawk);
-    }
-
-    if (mm->callsign_valid) {
-        printf("  Ident:         %s\n",
-                mm->callsign);
-    }
-
-    if (mm->category_valid) {
-        printf("  Category:      %02X\n",
-                mm->category);
-    }
-
-    if (mm->cpr_valid) {
-        printf("  CPR type:      %s\n"
-                "  CPR odd flag:  %s\n",
-                cpr_type_to_string(mm->cpr_type),
-                mm->cpr_odd ? "odd" : "even");
-
-        if (mm->cpr_decoded) {
-            printf("  CPR latitude:  %.5f (%u)\n"
-                    "  CPR longitude: %.5f (%u)\n"
-                    "  CPR decoding:  %s\n"
-                    "  NIC:           %u\n"
-                    "  Rc:            %.3f km / %.1f NM\n",
-                    mm->decoded_lat,
-                    mm->cpr_lat,
-                    mm->decoded_lon,
-                    mm->cpr_lon,
-                    mm->cpr_relative ? "local" : "global",
-                    mm->decoded_nic,
-                    mm->decoded_rc / 1000.0,
-                    mm->decoded_rc / 1852.0);
-        } else {
-            printf("  CPR latitude:  (%u)\n"
-                    "  CPR longitude: (%u)\n"
-                    "  CPR decoding:  none\n",
-                    mm->cpr_lat,
-                    mm->cpr_lon);
-        }
-    }
-
-    if (mm->accuracy.nic_a_valid) {
-        printf("  NIC-A:         %d\n", mm->accuracy.nic_a);
-    }
-    if (mm->accuracy.nic_b_valid) {
-        printf("  NIC-B:         %d\n", mm->accuracy.nic_b);
-    }
-    if (mm->accuracy.nic_c_valid) {
-        printf("  NIC-C:         %d\n", mm->accuracy.nic_c);
-    }
-    if (mm->accuracy.nic_baro_valid) {
-        printf("  NIC-baro:      %d\n", mm->accuracy.nic_baro);
-    }
-    if (mm->accuracy.nac_p_valid) {
-        printf("  NACp:          %d\n", mm->accuracy.nac_p);
-    }
-    if (mm->accuracy.nac_v_valid) {
-        printf("  NACv:          %d\n", mm->accuracy.nac_v);
-    }
-    if (mm->accuracy.gva_valid) {
-        printf("  GVA:           %d\n", mm->accuracy.gva);
-    }
-    if (mm->accuracy.sil_type != AIRCRAFT_META__SIL_TYPE__SIL_INVALID) {
-        const char *sil_description;
-        switch (mm->accuracy.sil) {
-            case 1:
-                sil_description = "p <= 0.1%";
-                break;
-            case 2:
-                sil_description = "p <= 0.001%";
-                break;
-            case 3:
-                sil_description = "p <= 0.00001%";
-                break;
-            default:
-                sil_description = "p > 0.1%";
-                break;
-        }
-        printf("  SIL:           %d (%s, %s)\n",
-                mm->accuracy.sil,
-                sil_description,
-                sil_type_to_string(mm->accuracy.sil_type));
-    }
-    if (mm->accuracy.sda_valid) {
-        printf("  SDA:           %d\n", mm->accuracy.sda);
-    }
-
-    if (mm->opstatus.valid) {
-        printf("  Aircraft Operational Status:\n");
-        printf("    Version:            %d\n", mm->opstatus.version);
-
-        printf("    Capability classes: ");
-        if (mm->opstatus.cc_acas) printf("ACAS ");
-        if (mm->opstatus.cc_cdti) printf("CDTI ");
-        if (mm->opstatus.cc_1090_in) printf("1090IN ");
-        if (mm->opstatus.cc_arv) printf("ARV ");
-        if (mm->opstatus.cc_ts) printf("TS ");
-        if (mm->opstatus.cc_tc) printf("TC=%d ", mm->opstatus.cc_tc);
-        if (mm->opstatus.cc_uat_in) printf("UATIN ");
-        if (mm->opstatus.cc_poa) printf("POA ");
-        if (mm->opstatus.cc_b2_low) printf("B2-LOW ");
-        if (mm->opstatus.cc_lw_valid) printf("L/W=%d ", mm->opstatus.cc_lw);
-        if (mm->opstatus.cc_antenna_offset) printf("GPS-OFFSET=%d ", mm->opstatus.cc_antenna_offset);
-        printf("\n");
-
-        printf("    Operational modes:  ");
-        if (mm->opstatus.om_acas_ra) printf("ACASRA ");
-        if (mm->opstatus.om_ident) printf("IDENT ");
-        if (mm->opstatus.om_atc) printf("ATC ");
-        if (mm->opstatus.om_saf) printf("SAF ");
-        printf("\n");
-
-        if (mm->mesub == 1)
-            printf("    Track/heading:      %s\n", heading_type_to_string(mm->opstatus.tah));
-        printf("    Heading ref dir:    %s\n", heading_type_to_string(mm->opstatus.hrd));
-    }
-
-    if (mm->nav.heading_valid)
-        printf("  Selected heading:        %.1f\n", mm->nav.heading);
-    if (mm->nav.fms_altitude_valid)
-        printf("  FMS selected altitude:   %u ft\n", mm->nav.fms_altitude);
-    if (mm->nav.mcp_altitude_valid)
-        printf("  MCP selected altitude:   %u ft\n", mm->nav.mcp_altitude);
-    if (mm->nav.qnh_valid)
-        printf("  QNH:                     %.1f millibars\n", mm->nav.qnh);
-    if (mm->nav.altitude_source != NAV_ALT_INVALID) {
-        printf("  Target altitude source:  ");
-        switch (mm->nav.altitude_source) {
-            case NAV_ALT_AIRCRAFT:
-                printf("aircraft altitude\n");
-                break;
-            case NAV_ALT_MCP:
-                printf("MCP selected altitude\n");
-                break;
-            case NAV_ALT_FMS:
-                printf("FMS selected altitude\n");
-                break;
-            default:
-                printf("unknown\n");
-        }
-    }
-
-    if (mm->nav.modes_valid) {
-        printf("  Nav modes:               %s\n", nav_modes_to_string(mm->nav.modes));
-    }
-
-    if (mm->emergency_valid) {
-        printf("  Emergency/priority:      %s\n", emergency_to_string(mm->emergency));
-    }
-
-    printf("\n");
-    fflush(stdout);
-}
-
-//
-//=========================================================================
-//
-// When a new message is available, because it was decoded from the RTL device,
-// file, or received in the TCP input port, or any other way we can receive a
-// decoded message, we call this function in order to use the message.
-//
-// Basically this function passes a raw message to the upper layers for further
-// processing and visualization
-//
-
-void useModesMessage(struct modesMessage *mm) {
-    struct aircraft *a;
-
-    ++Modes.stats_current.messages_total;
-
-    // Track aircraft state
-    a = trackUpdateFromMessage(mm);
-
-    // In non-interactive non-quiet mode, display messages on standard output
-    if (!Modes.interactive && !Modes.quiet && (!Modes.show_only || mm->addr == Modes.show_only) && !mm->sbs_in) {
-        displayModesMessage(mm);
-    }
-
-    // Feed output clients.
-    // If in --net-verbatim mode, do this for all messages.
-    // Otherwise, apply a sanity-check filter and only
-    // forward messages when we have seen two of them.
-
-    if (Modes.net && !mm->sbs_in) {
-        if (Modes.net_verbatim || mm->msgtype == 32 || !a) {
-            // Unconditionally send
-            modesQueueOutput(mm, a);
-        } else if (a->meta.messages > 1) {
-            // Suppress the first message. When we receive a second message,
-            // emit the first two messages.
-            if (a->meta.messages == 2) {
-                modesQueueOutput(&a->first_message, a);
-            }
-            modesQueueOutput(mm, a);
-        }
     }
 }
